Filepath Processing & Filter by Extensions
Added additional text processing for entered filepaths so that " marks will be processed and ignored. Copying filepaths is not easier and doesn't require user interaction to strip them from entered text. Added an option to define the filter extension that is used when searching for files within a directory on encryption. This allows for refinement of the files that are grabbed. Also added the ability to clear the console with a command