Removed Async Functionality
The overhead of async was contributing to the out of memory and decryption failure (Not sure how/why). Removed this as the application is a console application so it's not highly important. For whatever reason something was occurring that was resulting in the encryption/decryption process to be modified. Without the async, managed to encrypt and decrypt a collection of 2049 relatively small text files. Will probably need to look into what causes memory to not be cleared despite falling out of scope and dispose functions being called